pyproject.toml
[project]
name = "agentic-voice-assistant"
version = "0.1.0"
description = "Voice-to-Voice AI Assistant for Product Discovery using LangGraph, MCP and RAG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.5",
    "datasets>=4.4.1",
    "groq>=0.36.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "kagglehub>=0.3.13",
    "langgraph>=1.0.3",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "streamlit>=1.51.0",
    "sentence-transformers>=5.1.2",
    "fastapi>=0.124.0",
    "uvicorn>=0.38.0",
    "torch>=2.9.1",
]

[tool.uv]
dev-dependencies = ["pre-commit", "ruff"]
